Rename css loader file
