+++ packages/marko-web/express/index.js
@@ -14,7 +14,7 @@ const loadDocument = require('./load-document');
 const oembed = require('./oembed');
 const cdn = require('./cdn');
 const assetLoader = require('./asset-loader');
-const cssLoader = require('./css');
+const cssLoader = require('./css-loader');
 const rss = require('./rss');
 const sitemaps = require('./sitemaps');
 const { version } = require('../package.json');
